initial patches code
